requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aniso8601==7.0.0          # via graphene
certifi==2019.9.11        # via requests, sentry-sdk
chardet==3.0.4            # via requests
django-cors-headers==3.1.1
django-environ==0.4.5
django-helusers==0.5.2
django==2.2.6
djangorestframework==3.10.3  # via drf-oidc-auth
drf-oidc-auth==0.9        # via django-helusers
ecdsa==0.13.3             # via python-jose
future==0.17.1            # via pyjwkest, python-jose
graphene-django==2.6.0
graphene==2.1.8           # via graphene-django
graphql-core==2.2.1       # via graphene, graphene-django, graphql-relay
graphql-relay==2.0.0      # via graphene
idna==2.8                 # via requests
promise==2.2.1            # via graphene-django, graphql-core, graphql-relay
psycopg2==2.8.3
pyasn1==0.4.7             # via rsa
pycryptodomex==3.9.0      # via pyjwkest
pyjwkest==1.4.2           # via drf-oidc-auth
python-jose==3.0.1        # via django-helusers
pytz==2019.3              # via django
requests==2.22.0          # via django-helusers, pyjwkest
rsa==4.0                  # via python-jose
rx==1.6.1                 # via graphql-core
sentry-sdk==0.12.3
singledispatch==3.4.0.3   # via graphene-django
six==1.12.0               # via graphene, graphene-django, graphql-core, graphql-relay, promise, pyjwkest, python-jose, singledispatch
sqlparse==0.3.0           # via django
urllib3==1.25.6           # via requests, sentry-sdk
